Move tasks directory into src folder
