+++ server/src/app.module.ts
@@ -4,7 +4,7 @@ import { AuthModule } from './routes/auth/auth.module';
 import config from './config';
 import { UsersModule } from './routes/user/users.module';
 import { ScheduleModule } from '@nestjs/schedule';
-import { TasksModule } from './common/tasks/tasks.module';
+import { TasksModule } from './tasks/tasks.module';
 import { UserSeederModule } from './database/seeders/user/user-seeder.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
