+++ server/src/tasks/tasks.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UsersModule } from '../../routes/user/users.module';
+import { UsersModule } from '../routes/user/users.module';
 import { TasksService } from './tasks.service';
 
 @Module({
